# paperless-backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Paperless REST API
  description: |
    Document Management System REST API for the SWEN3 Paperless project.
    Provides endpoints for:
    - Document Upload: file storage, automated OCR processing, Elasticsearch full-text indexing, and GenAI summarization.
    - Document Search: full-text and fuzzy search powered by Elasticsearch.
    - Document Management: document lifecycle (details, metadata update, deletion, download) and metadata (document types).
    - User and Team Management: authentication (login, me), user management (User-MM), teams, and role-based memberships (User <-- m:n (role) --> Team; roles: ADMIN, READONLY, READ_WRITE).
  version: 1.0.0
  x-api-id: d0184f38-b98d-11e7-9c56-68f728c1ba70
  x-audience: company-internal
  contact:
    name: Philip Duong & Patrick Hornek
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local Docker / Development Server
security:
  - bearerAuth: []
...
paths:
...
  /documents:
    get:
      tags:
        - Documents
      summary: Search and list documents
      description: Returns a paginated list of documents visible to the authenticated user, with optional metadata filtering.
      operationId: searchDocuments
      parameters:
        - name: page
          in: query
          description: Zero-based page index.
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned.
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          description: Sorting criteria in the format property,(asc|desc).
          required: false
          schema:
            type: string
            default: created_at,desc
        - name: document_type_id
          in: query
          description: Filter by Document Type ID.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved paged documents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPageResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Documents
      summary: Upload a new document
      description: |
        Uploads a PDF document to file storage (MinIO). If team_id is absent, the document is private and visible only to its uploader. Sharing with a team requires current READ_WRITE or ADMIN membership in that team; insufficient access returns 403 and a nonexistent team returns 404.
        Automatically triggers asynchronous processing:
        - Performs OCR recognition to extract document text.
        - Automatically indexes extracted text for full-text search in Elasticsearch.
        - Automatically generates an AI summary using Google Gemini.
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - document
                - title
              properties:
                document:
                  type: string
                  format: binary
                  description: The binary PDF document file.
                title:
                  type: string
                  description: Required nonblank title for the uploaded document.
                  example: Electricity Invoice October 2026
                  minLength: 1
                  maxLength: 255
                document_type_id:
                  type: string
                  format: uuid
                  nullable: true
                  example: "550e8400-e29b-41d4-a716-446655440004"
                team_id:
                  type: string
                  format: uuid
                  description: Optional team to share the document with; omit to keep it private. The uploader must have READ_WRITE or ADMIN access to the team.
                  example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        '201':
          description: Document successfully uploaded and queued for processing.
          headers:
            Location:
              description: URI of the newly uploaded document resource.
              schema:
                type: string
                format: uri
                example: "/documents/550e8400-e29b-41d4-a716-446655440002"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document details by ID
      description: Returns full document metadata only when the authenticated user can access it; private documents are uploader-only and team documents are visible only to current members. A removed uploader loses access and receives 404.
      operationId: getDocumentById
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Document details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Documents
      summary: Update document metadata
      description: |
        Any current writer may update document metadata (title and document type). Changing team_id or clearing team sharing is restricted to the uploader. Moving to another team requires write access to both the current source and destination teams; making the document private requires write access to the current source team. Omitting team_id or setting it to null does not clear sharing; set clear_team to true without team_id to make the document private. Supplying team_id together with clear_team: true is invalid and returns 400. A missing or inaccessible document, or a nonexistent destination team, returns 404; insufficient permission returns 403. A team-sharing transition that conflicts with the document's current sharing state returns 409.
      operationId: updateDocumentMetadata
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Documents
      summary: Delete a document
      description: Deletes a document record, file from MinIO, and Elasticsearch search index entry; only the private uploader or a current READ_WRITE/ADMIN team member may delete it. Insufficient write access returns 403; a missing or inaccessible document returns 404.
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Document successfully deleted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /documents/{id}/download:
    get:
      tags:
        - Documents
      summary: Download original document file
      description: Streams the original uploaded PDF file from object storage (MinIO) when the authenticated user can access the document; private documents are uploader-only and team documents are visible only to current members. A missing or inaccessible document returns 404.
      operationId: downloadDocumentFile
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Binary PDF file content.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # SEARCH (ELASTICSEARCH)
  # ============================================================================
...
  /document-types:
    get:
      tags:
        - Document Types
      summary: List all document types
      operationId: searchDocumentTypes
      responses:
        '200':
          description: List of document types.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Document Types
      summary: Create a new document type
      operationId: createDocumentType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentTypeRequest'
      responses:
        '201':
          description: Document type created successfully.
          headers:
            Location:
              description: URI of the newly created document type resource.
              schema:
                type: string
                format: uri
                example: "/document-types/550e8400-e29b-41d4-a716-446655440004"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /document-types/{id}:
    get:
      tags:
        - Document Types
      summary: Get document type by ID
      operationId: getDocumentTypeById
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Document type retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Document Types
      summary: Update a document type
      operationId: updateDocumentType
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentTypeRequest'
      responses:
        '200':
          description: Document type updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Document Types
      summary: Delete a document type
      operationId: deleteDocumentType
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Document type deleted successfully.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ==============================================================================
# COMPONENTS & SCHEMAS
# ==============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Standard Authorization Bearer header using an opaque access token.
  parameters:
    PathId:
      name: id
      in: path
      description: Unique identifier of the resource.
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    PageMetadata:
      type: object
      required:
        - page
        - size
        - total_elements
        - total_pages
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        total_elements:
          type: integer
          format: int64
          example: 25
        total_pages:
          type: integer
          example: 3

    # --- Role Enum ---
...
    ProcessingStatus:
      type: string
      enum:
        - PENDING
        - OCR_IN_PROGRESS
        - GENAI_IN_PROGRESS
        - COMPLETED
        - FAILED
      example: COMPLETED
    DocumentResponse:
      type: object
      required:
        - id
        - owner_id
        - title
        - original_filename
        - content_type
        - file_size
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        owner_id:
          type: string
          format: uuid
          readOnly: true
          description: User ID of the document uploader and owner.
          example: "550e8400-e29b-41d4-a716-446655440000"
        team_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the team currently sharing this document; null when private.
          example: "550e8400-e29b-41d4-a716-446655440001"
        title:
          type: string
          example: Electricity Invoice October 2026
        original_filename:
          type: string
          example: invoice_oct_2026.pdf
        content_type:
          type: string
          example: application/pdf
        file_size:
          type: integer
          format: int64
          description: Size in bytes.
          example: 245760
        status:
          $ref: '#/components/schemas/ProcessingStatus'
        ocr_content:
          type: string
          nullable: true
          description: Extracted text from OCR worker.
          example: "INVOICE #9821\nAmount Due: 142.50 EUR\nDate: 2026-10-01"
        summary:
          type: string
          nullable: true
          description: AI summary generated by Google Gemini worker.
          example: "Monthly utility invoice for electricity services totaling 142.50 EUR due by end of month."
        storage_key:
          type: string
          description: Object key in MinIO.
          example: "documents/2026/10/1_invoice_oct_2026.pdf"
        document_type:
          $ref: '#/components/schemas/DocumentTypeResponse'
          nullable: true
        created_at:
          type: string
          format: date-time
          example: "2026-10-15T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-10-15T14:32:10Z"
    DocumentPageResponse:
      type: object
      required:
        - pagination
        - items
      properties:
        pagination:
          $ref: '#/components/schemas/PageMetadata'
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
    UpdateDocumentRequest:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
          nullable: true
          description: Destination team ID. Omitting this field or setting it to null leaves current sharing unchanged; null alone does not make the document private.
          example: "550e8400-e29b-41d4-a716-446655440001"
        clear_team:
          type: boolean
          default: false
          description: Set true without team_id to remove team sharing and make the document private.
        title:
          type: string
          description: Optional nonblank replacement document title.
          minLength: 1
          maxLength: 255
          example: "Updated Invoice Title"
        document_type_id:
          type: string
          format: uuid
          nullable: true
          example: "550e8400-e29b-41d4-a716-446655440004"

    # --- Search Result Schemas ---
...
    DocumentTypeResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440004"
        name:
          type: string
          example: Invoice
        description:
          type: string
          nullable: true
          example: Billing and payment invoices
    DocumentTypeListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentTypeResponse'
    CreateDocumentTypeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Required nonblank document type name.
          minLength: 1
          maxLength: 100
          example: Contract
        description:
          type: string
          nullable: true
          example: Employment, tenancy, and service contracts
    UpdateDocumentTypeRequest:
      type: object
      properties:
        name:
          type: string
          description: Optional nonblank replacement document type name.
          minLength: 1
          maxLength: 100
          example: Legal Contract
        description:
          type: string
          nullable: true
          example: Legal and financial binding contracts

    # --- RFC 7807 Problem JSON ---
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: "urn:problem-type:not-found"
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: "Resource Not Found"
        status:
          type: integer
          description: The HTTP status code generated by the origin server.
          example: 404
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: "Document with id 550e8400-e29b-41d4-a716-446655440002 was not found."
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
          example: "/documents/550e8400-e29b-41d4-a716-446655440002"
        invalid_params:
          type: array
          description: An array of invalid parameter errors (RFC 7807 extension).
          items:
            $ref: '#/components/schemas/InvalidParam'
    InvalidParam:
      type: object
      required:
        - name
        - reason
      properties:
        name:
          type: string
          description: The name of the invalid parameter or field.
          example: "title"
        reason:
          type: string
          description: Explanation of why the parameter is invalid.
          example: "Title must not be blank."

  # ============================================================================
  # RESPONSES
  # ============================================================================
  responses:
    BadRequestError:
      description: Bad Request - Validation failure or malformed payload.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnauthorizedError:
      description: Unauthorized - Authentication is required or credentials are invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    ForbiddenError:
      description: Forbidden - The authenticated user is not authorized to perform this operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFoundError:
      description: Not Found - The requested entity does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    ConflictError:
      description: Conflict - A resource with the same identifier or unique constraint already exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: Internal Server Error - Unexpected failure on the server.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'